Use JsonTextWriter (instead of JObject) and extract out common code by using UnifiedJsonWriter extract out common code by using UnifiedJsonWriterr
