+++ src/managed/Microsoft.Extensions.DependencyModel/Microsoft.Extensions.DependencyModel.csproj
@@ -32,6 +32,7 @@
         <Compile Remove="DependencyContextJsonReader.JsonTextReader.cs" />
         <Compile Remove="DependencyContextWriter.JsonTextWriter.cs" />
         <Compile Remove="UnifiedJsonReader.JsonTextReader.cs" />
+        <Compile Remove="UnifiedJsonWriter.JsonTextWriter.cs" />
       </ItemGroup>
     </When>
     <Otherwise>
@@ -43,6 +44,7 @@
         <Compile Remove="DependencyContextJsonReader.Utf8JsonReader.cs" />
         <Compile Remove="DependencyContextWriter.Utf8JsonWriter.cs" />
         <Compile Remove="UnifiedJsonReader.Utf8JsonReader.cs" />
+        <Compile Remove="UnifiedJsonWriter.Utf8JsonWriter.cs" />
       </ItemGroup>
     </Otherwise>
   </Choose>
